--- screens/overview.tsx
@@ -1,34 +1,0 @@
-import { useNavigation } from '@react-navigation/native';
-import { StackNavigationProp } from '@react-navigation/stack';
-import { ScreenContent } from 'components/ScreenContent';
-import { StyleSheet, View } from 'react-native';
-
-import { Button } from '../components/Button';
-import { RootStackParamList } from '../navigation';
-
-type OverviewScreenNavigationProps = StackNavigationProp<RootStackParamList, 'Overview'>;
-
-export default function Overview() {
-  const navigation = useNavigation<OverviewScreenNavigationProps>();
-
-  return (
-    <View style={styles.container}>
-      <ScreenContent path="screens/overview.tsx" title="Overview" />
-      <Button
-        onPress={() =>
-          navigation.navigate('Details', {
-            name: 'Dan',
-          })
-        }
-        title="Show Details"
-      />
-    </View>
-  );
-}
-
-export const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 24,
-  },
-});
