+++ src/screens/Inicio/index.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { SafeAreaView } from 'react-native'
+
+export function Inicio() {
+    return (
+        <SafeAreaView>
+            
+        </SafeAreaView>
+    )
+}
