+++ src/screens/Login/index.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { SafeAreaView } from 'react-native'
+
+export function Login() {
+    return (
+        <SafeAreaView>
+            
+        </SafeAreaView>
+    )
+}
